Add Metadata to event for Unit Tests
When creating an event, Broadway gives you the possibility to enrich the
DomainMessage with metadata. This is done via Decorators and Enrichers.
However, in Unit Testing there is no way to mock metadata because no
Decorators is called.
To mock this behavior for testing, just add a new parameter to the when()-method which gives the possibility to
pass the Metadata further down to the DomainMessage.

Background:
In some cases it is necessary to access the stored metadata when
applying the Events onto your ReadModel. If the stored data is not
available in testing context, the tests fails with an "Index not found"
exception.